+++ packages/cli-node/package.json
@@ -35,9 +35,13 @@
     "@backstage/errors": "workspace:^",
     "@backstage/types": "workspace:^",
     "@manypkg/get-packages": "^1.1.3",
+    "@yarnpkg/lockfile": "^1.1.0",
     "@yarnpkg/parsers": "^3.0.0",
+    "chalk": "^4.0.0",
     "fs-extra": "^11.2.0",
+    "ora": "^5.3.0",
     "semver": "^7.5.3",
+    "yaml": "^2.0.0",
     "zod": "^3.22.4"
   },
   "devDependencies": {
