+++ packages/cli-node/src/errors.ts
