+++ packages/cli-node/src/index.ts
@@ -24,3 +24,5 @@ export * from './git';
 export * from './monorepo';
 export * from './roles';
 export * from './pacman';
+// TODO probably not needed
+export * from './run';
